==> analysis/1. Palette pipeline (Drosophila slices as example).Rmd <==
---
title: "Inferring spatial gene expression patterns from _Drosophila_ pseudo bulk data"
author: "Yang Dong"
date: "2024-03-26"
output:
  html_document:
    toc: yes
    toc_float: yes
    code_folding: hide
editor_options:
  chunk_output_type: console
---
<style type="text/css">
  body{
  font-size: 12pt;
}
</style>

### Setup
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE)
```

```{r packages}
library(limma)
library(Seurat)
library(readr)
library(tidyverse)
library(magrittr)
library(pander)
library(SingleCellExperiment)
library(reshape2)
library(BayesSpace)
library(MuSiC)
library(ggplot2)
library(ggpubr)
library(Matrix.utils)

theme_set(theme_bw())
panderOptions("big.mark", ",")
panderOptions("table.split.table", Inf)
panderOptions("table.style", "rmarkdown")
if (interactive()) setwd(here::here("analysis"))
```

### Summary
<p align="justify">Here we take 2 neighboring ST slices (slice 4 and slice 5) from _Drosophila_ ST data. Slice 4 will be used as the ST reference, while slice 5 will be converted a pseudo bulk data. Palette will then be implemented to infer the expression patterns from the pseudo bulk data using the slice 4 as the ST reference. The inferred expression patterns will then be compared with the real expression of slice 5 ST data.  
</p>

### Load data
Load the neighboring _Drosophila_ ST slice.
```{r load data}
Drosophila_slice <- readRDS(here::here("test","Drosophila_slice.rds"))
```

### _Drosophila_ ST slices 
Here we visualize the ST slice and each ST spot is colored with spot annotation. The two ST slices show similar cell type distribution. 
```{r Drosophila ST slices, out.width="60%", fig.align = 'center',fig.cap = "The adjacent _Drosophila_ ST slices."}
# Assign color to each cell type
color_df <- data.frame(unique(Drosophila_slice$annotation))
colnames(color_df)[1] <- "annotation"
color_df$color <- c("#FFF799","#2ED9FF","#683b79","#009200","#6e9BC5","#F091A0","#FEAF16","#ff66ff","#AEF359","#C0C0C0")
x <- Drosophila_slice@meta.data %>%
  left_join(color_df)
Drosophila_slice@meta.data$color <- x$color

slice_4 <- subset(Drosophila_slice, subset = slice_ID == "E14-16h_a_S04")
slice_5 <- subset(Drosophila_slice, subset = slice_ID == "E14-16h_a_S05")

# Visualize ST slice
g1 <- ggplot(slice_4@meta.data %>% as.data.frame(),aes(x=new_x, y=new_y, color=color)) +
  geom_point(size=0.5) +
  theme_classic() +
  xlim(-25, 25) +
  ylim(-25, 25) +
    coord_fixed(ratio = 1, xlim = NULL, ylim = NULL, expand = TRUE, clip = "on") +
  scale_color_manual(values = levels(slice_4@meta.data$color %>% as.factor())) +
  theme_classic() +
  theme(axis.text.x=element_blank(),
        axis.text.y=element_blank(),
        axis.title.x=element_blank(),
        axis.title.y=element_blank(),
        legend.position="none")
g2 <- ggplot(slice_5@meta.data %>% as.data.frame(),aes(x=new_x, y=new_y, color=color)) +
  geom_point(size=0.5) +
  theme_classic() +
  xlim(-25, 25) +
  ylim(-25, 25) +
    coord_fixed(ratio = 1, xlim = NULL, ylim = NULL, expand = TRUE, clip = "on") +
  scale_color_manual(values = levels(slice_5@meta.data$color %>% as.factor())) +
  theme_classic() +
  theme(axis.text.x=element_blank(),
        axis.text.y=element_blank(),
        axis.title.x=element_blank(),
        axis.title.y=element_blank(),
        legend.position="none")

ggarrange(g1, g2, ncol = 2)
```
<p>&nbsp;</p>
### Construct pseudo bulk data and select highly expressed genes
<p align="justify">Aggregate expression of ST data to generate pseudo bulk. Slice 5 is convert to pseudo bulk, while the aggregated expression of ST data (slice 4) is used for selecting highly expressed genes.
</p>
```{r}
# Convert both to pseudo bulk data
slice_sce <- SingleCellExperiment(assays = list(counts = Drosophila_slice@assays$RNA@counts), 
                           colData = Drosophila_slice@meta.data)
groups <- colData(slice_sce)[, c("slice_ID")]
slice_bulk <- aggregate.Matrix(t(counts(slice_sce)), 
                       groupings = groups, fun = "sum") %>% 
  as.data.frame() %>% t()

# Slice 5 as pseudo bulk
slice_5_bulk <- slice_bulk[,2, drop = FALSE] %>% set_colnames("slice_5")
# Slice 4 as aggregated ST data for selecting highly expressed genes
slice_4_ST_sum <- slice_bulk[,1, drop = FALSE] %>% set_colnames("slice_4_ST")

# Select highly expressed genes in ST data
slice_4_ST_sum_high_exp <- slice_4_ST_sum %>%
  as.data.frame() %>%
  dplyr::filter(rownames(.) %in% rownames(slice_5_bulk)) %>%
  dplyr::filter(.$slice_4_ST > 30)

# Reorder genes of pseudo bulk data (slice 5) based on the ST data (slice 4)
slice_5_bulk_reorder <- slice_5_bulk[match(rownames(slice_4_ST_sum_high_exp), rownames(slice_5_bulk)),,drop = FALSE] %>%
  as.data.frame()

# Define the ratio of the pseudo bulk (slice 5) and the aggregate ST (slice 4) as gene filter
gene_filter <- (slice_5_bulk_reorder/slice_4_ST_sum_high_exp) %>%
  set_colnames("ratio") %>%
  .[is.finite(rowSums(.)),, drop = FALSE]
```

<p align="justify">We use density plot to visualize the expression ratios of the pseudo bulk (slice 5) and the aggregated ST (slice 4). We take the top 2000 genes closest to the mean and median ratios for the following analyses. These genes are highly expressed in both pseudo bulk and ST data.
</p>
```{r, out.width="40%", fig.align = 'center',fig.cap = "Density plot showing the expression ratio of pseudo bulk and aggregated ST data."}
ggplot(gene_filter, aes(x=ratio)) + 
  geom_density() + theme_classic() +
  theme(text = element_text(size = 20)) 
  
# The difference of gene ratio to the mean and median ratio
gene_filter$differ <- abs(gene_filter$ratio - (mean(gene_filter$ratio)+median(gene_filter$ratio))/2)

# Select the top 2000 genes closer to the mean and median for spatial clustering and deconvolution
high_exp_genes <- gene_filter %>%
  .[order(.$differ),] %>%
   top_n(.,-2000)
```
<p>&nbsp;</p>
### Spatial clustering of ST data
<p align="justify">BayesSpace is employed for spatial clustering on ST data (slice 4). Only the selected 2000 highly expressed genes are used for spatial clustering. The cluster identities are then assigned to each ST spot. 
</p>
```{r, out.width="40%", fig.align = 'center', fig.cap = "Spatial Clustering of ST data."}
# Convert coordinates into pixel coordinates for BayesSpace
metadata <-  slice_4@meta.data
metadata$col <-  metadata$new_x %>% as.numeric() %>% as.integer()
metadata$row <-  metadata$new_y %>% as.numeric() %>% as.integer()

# Construct sce object
sce <- SingleCellExperiment(assays = list(counts = slice_4@assays$RNA@counts, logcounts = slice_4@assays$RNA@data),
                            colData=metadata)

# Only take highly expressed genes for spatial clustering
sce_filtered <- sce[rownames(high_exp_genes),]
sce_filtered <- spatialPreprocess(sce_filtered, platform="ST",
                              n.PCs=4,n.HVGs = 2000, log.normalize=TRUE)
# Select the number of clusters
p <- qTune(sce_filtered, qs=seq(2, 10), platform="ST")

# Clustering with BayesSpace
set.seed(149)
spatial_cluster <- spatialCluster(p, q=5, platform="ST", d=7,
                           init.method="kmeans", model="t", gamma=2,
                           nrep=10000, burn.in=100,
                           save.chain=TRUE)

# Visualizing spatial clusters
clusterPlot(spatial_cluster, color="black") +
  theme_bw() +
  xlab("") +
  ylab("") + 
  theme(text = element_text(size = 20)) 

slice_4$spatial_cluster <- spatial_cluster$spatial.cluster
Idents(slice_4) <- slice_4$spatial_cluster

# Check the proportions of each cluster
prop_table <- prop.table(table(slice_4$spatial_cluster)) %>% t()
colnames(prop_table) <- c("cluster 1","cluster 2","cluster 3","cluster 4","cluster 5")
rownames(prop_table) <- "Proportion"
knitr::kable(prop_table,
             caption = "The proportions of each cluster.")
```
<p>&nbsp;</p>
### Pseudo bulk deconvolution
<p align="justify">MuSiC is employed for deconvolution. Also, only the selected 2000 highly expressed genes are used for deconvolution. We can then achieve the abundances of each cluster in pseudo bulk data.
</p>
```{r Use MuSiC for deconvolution}
# Pre-process data for MuSiC
bulk.est <- Biobase::ExpressionSet(assayData = slice_bulk %>% as.matrix())
sc.eset <- ExpressionSet(assayData = as.matrix(GetAssayData(slice_4[rownames(high_exp_genes),])), phenoData = new("AnnotatedDataFrame", slice_4@meta.data))

slice_4$sample_id <- colnames(slice_4)
sce <- SingleCellExperiment(assays = list(counts = slice_4[rownames(high_exp_genes),]@assays$RNA@counts, logcounts = slice_4[rownames(high_exp_genes),]@assays$RNA@data),
                            colData=slice_4@meta.data)

# MuSiC deconvolution
Est.pro = music_prop(bulk.mtx = exprs(bulk.est), sc.sce = sce,  clusters = "spatial_cluster", samples = "sample_id")
proportion <- Est.pro$Est.prop.weighted
proportion_reorder <- proportion[, c("1","2","3","4","5")]

# The estimated proportion of each cluster in pseudo bulk data
prop_table_2 <- proportion_reorder[2,] %>% t()
colnames(prop_table_2) <- c("cluster 1","cluster 2","cluster 3","cluster 4","cluster 5")
rownames(prop_table_2) <- "Proportion"
knitr::kable(prop_table_2,
             caption = "The estimated proportions of each cluster in pseudo bulk data (slice 5).")
```
<p>&nbsp;</p>
### Variable factor calculation
<p align="justify">Here we define a variable factor to adjust the expression difference between pseudo bulk and ST data. The variable factor is defined as the ratio of bulk expression matrix to the pseudo bulk matrix. Each gene has its own variable factor to adjust its expression.
</p>
```{r Variable factor calculation}
# The average gene expression of each cluster
cluster.averages <- AverageExpression(slice_4) %>%
  as.data.frame() %>%
  as.matrix()

cluster.averages_filtered <- cluster.averages %>%
  as.data.frame() %>%
  rownames_to_column() %>%
  dplyr::filter(.$rowname %in% rownames(slice_bulk)) %>%
  column_to_rownames(var= "rowname")  %>%
  as.matrix() %>%
  .[,c("RNA.1","RNA.2","RNA.3","RNA.4","RNA.5")]

# The abundance of each cluster in pseudo bulk 
abundance <- proportion_reorder %>%
  .[2,,drop=FALSE] %>%
  as.data.frame() %>%
  as.matrix()


slice_5_bulk_reorder <- slice_5_bulk[match(rownames(cluster.averages_filtered), rownames(slice_5_bulk)),,drop=FALSE] %>%
  as.data.frame()

# Define variable factor
variable_factor <- (slice_5_bulk_reorder/(cluster.averages_filtered %*% (abundance %>% t()))) %>%
  .[is.finite(rowSums(.)),, drop = FALSE] %>%
  as.data.frame() %>%
  set_colnames("var")

variable_factor_matrix <- rep(variable_factor,5) %>%
  as.data.frame() %>%
  set_colnames(c("1","2","3","4","5"))
rownames(variable_factor_matrix) <- rownames(variable_factor)
```

```{r print variable factor, echo=FALSE}
print("The top 5 rows of variable factor")
head(variable_factor,5)
# Use variable factor to adjust the average gene expression of each cluster and achieve the adjusted overall gene expression of each cluster called adjusted_matrix
```

<p align="justify">The adjusted matrix is the overall gene expression of each cluster, which is obtained by taking the dot product of the cluster expression matrix of ST slice and the variable factor. 
</p>
```{r adjusted matrix}
adjusted_matrix <- ((cluster.averages_filtered[rownames(variable_factor_matrix),] * variable_factor_matrix)) 
colnames(adjusted_matrix) <- c("cluster 1","cluster 2","cluster 3","cluster 4","cluster 5")
```

```{r print adjusted matrix, echo=FALSE}
print("The top 5 rows of adjusted matrix")
head(adjusted_matrix,5)
```

<p>&nbsp;</p>
### Estimate expression of each spot
<p align="justify">
The expression of each spot is estimated through a loop algorithm. In each interaction of the loop, the procedure begins by selecting one random spot and its neighboring spots. The number of neighboring spots can be changed depending on the data. The expression of spots belonging to the same cluster is aggregated to form a pseudo bulk data called regional cluster ST (RST).
<p>&nbsp;</p>
Here we assume that the ratio of RST to the entire ST data is equal to the ratio of the adjusted RST to the adjusted matrix. We define the ratio of RST to the entire ST as regional factor K. The adjust regional ST equals to the dot product of adjusted matrix and regional factor K. The expression of adjusted RST is then evenly allocated into the selected spots of this cluster.
After thousands of interactions, the average expression of each spot is almost stable and considered as the output estimated expression.
</p>
```{r expression allocation}
ST <- slice_4_ST_sum %>%
    as.data.frame() %>%
    rownames_to_column() %>%
    dplyr::select(Gene=rowname, Counts = 'slice_4_ST')


#### Nearest 8 spots
output_final <- data.frame(matrix(ncol = 0, nrow = length(rownames(variable_factor_matrix))))
rownames(output_final) <- rownames(variable_factor_matrix)
for (i in 1:5000) {
  # Select one random spot
  set.seed(i)
  slice_4.random.obj <- slice_4[, sample(colnames(slice_4), size = 1, replace=F)]
  spot <- slice_4.random.obj@meta.data[,c("new_x","new_y")] %>% as.data.frame()
  df <- slice_4@meta.data[,c("new_x","new_y")] %>% as.data.frame()
  # Calculate the distance from other spots to the selected spot
  df$distance = sqrt((df$new_x - spot$new_x)^2 + (df$new_y - spot$new_y)^2)
  # Here we select the nearest 8 spots. The number of the nearest spots can be changed here. 
  df <- df[order(df$distance),][1:9,]
  slice_4.random.neighbor <-  slice_4[,rownames(df)]
  output_region <- data.frame(matrix(ncol = 0, nrow = length(rownames(variable_factor_matrix))))
  rownames(output_region) <- rownames(variable_factor_matrix)
  # For each cluster in the selecting spots
  for (k in as.numeric(levels(as.factor(slice_4.random.neighbor$spatial_cluster)))){
    # Select the spots belonging to the same cluster
    cluster <- subset(slice_4.random.neighbor, subset = spatial_cluster == k)
    # Sum the expression of these spots from the same cluster
    RST <- colSums(t((cluster@assays$RNA@counts))) %>%
      as.data.frame() %>%
      rownames_to_column() %>%
      dplyr::select(Gene=rowname, Counts = '.')
    # Define a regional factor K as the ratio of RST to ST
    K <- RST$Counts/ST$Counts %>%
      as.data.frame()
    rownames(K) <- ST$Gene
    K_filtered <- K %>%
      rownames_to_column() %>%
      dplyr::select(Gene=rowname, factor = '.') %>%
      dplyr::filter(.$Gene %in% rownames(variable_factor_matrix)) %>%
      column_to_rownames(var= "Gene")  %>%
      as.matrix()
    # Adjusted regional bulk equals to the dot product of adjusted matrix and regional factor K
    region_bulk <- (adjusted_matrix[,k] %>% as.data.frame()) * K_filtered
    # Assign expression to each spot
    each_spot <- data.frame(region_bulk/length(cluster$orig.ident))
    output <- cbind(each_spot, rep(each_spot[1],  length(cluster$orig.ident)))
    output <- output[1:(length(output)-1)]
    colnames(output) <- rownames(cluster@meta.data)
    output_region <- output_region %>%
      cbind(output)
    }
  output_final <- output_final %>%
    cbind(output_region)
}

# Calculate the average expression of each spot
cell_list <- as.list(rownames(slice_4@meta.data))

spot_average_counts_final <- data.frame(matrix(ncol = 0, nrow = length(rownames(variable_factor_matrix))))
rownames(spot_average_counts_final) <- rownames(variable_factor_matrix)
spot_rep_times <- data.frame()

for ( i in 1:length(cell_list)) {
  spot <- colnames(output_final) == cell_list[i]
  spot_counts <- output_final[, spot, drop = FALSE]
  spot_rep_times <- spot_rep_times %>%
    rbind(length(colnames(spot_counts)))
  spot_average_counts <-  data.frame(rowSums(spot_counts))/length(spot_counts)
  colnames(spot_average_counts) <- cell_list[i]
  spot_average_counts_final <- spot_average_counts_final %>%
    cbind(spot_average_counts)
}

# Covert Palette-inferred data into seurat object
slice_4_adjusted <- CreateSeuratObject(spot_average_counts_final,assay = "RNA",meta.data = metadata)
slice_4_adjusted@images$image <- slice_4@images$image
saveRDS(slice_4_adjusted, here::here("test","slice_4_adjusted.rds"))
```
<p>&nbsp;</p>
### Using a loss function for a optimal iteration time (optional)
<p align="justify">
An optimal iteration condition can ensure the performance of the inference pipeline as well as refine computer efficiency. Rather than use a specific iteration time, we also incorporate a loss function in to the iteration step. Here we define the differences between the two matrices over a specific iteration interval as the loss, and then the iteration will be stopped when the differences between adjacent losses are below the threshold for 3 consecutive times.
</p>

```{r incorporate loss function}
# Set iteration interval as 500
check_interval <- 500 # Check for each 500 iterations
# The threshold for the differences
threshold <- 1 
max_iterations <- 10000
losses <- numeric(max_iterations)

loss_function <- function(mat1, mat2) {
  return(sqrt(sum((mat1 - mat2)^2)))
}

#### Nearest 8 spots
output_final <- data.frame(matrix(ncol = 0, nrow = length(rownames(variable_factor_matrix))))
rownames(output_final) <- rownames(variable_factor_matrix)
for (i in 1:max_iterations) {
  # Select one random spot
  set.seed(i)
  slice_4.random.obj <- slice_4[, sample(colnames(slice_4), size = 1, replace=F)]
  spot <- slice_4.random.obj@meta.data[,c("new_x","new_y")] %>% as.data.frame()
  df <- slice_4@meta.data[,c("new_x","new_y")] %>% as.data.frame()
  # Calculate the distance from other spots to the selected spot
  df$distance = sqrt((df$new_x - spot$new_x)^2 + (df$new_y - spot$new_y)^2)
  # Here we select the nearest 8 spots. The number of the nearest spots can be changed here. 
  df <- df[order(df$distance),][1:9,]
  slice_4.random.neighbor <-  slice_4[,rownames(df)]
  output_region <- data.frame(matrix(ncol = 0, nrow = length(rownames(variable_factor_matrix))))
  rownames(output_region) <- rownames(variable_factor_matrix)
  # For each cluster in the selecting spots
  for (k in as.numeric(levels(as.factor(slice_4.random.neighbor$spatial_cluster)))){
    # Select the spots belonging to the same cluster
    cluster <- subset(slice_4.random.neighbor, subset = spatial_cluster == k)
    # Sum the expression of these spots from the same cluster
    RST <- colSums(t((cluster@assays$RNA@counts))) %>%
      as.data.frame() %>%
      rownames_to_column() %>%
      dplyr::select(Gene=rowname, Counts = '.')
    # Define a regional factor K as the ratio of RST to ST
    K <- RST$Counts/ST$Counts %>%
      as.data.frame()
    rownames(K) <- ST$Gene
    K_filtered <- K %>%
      rownames_to_column() %>%
      dplyr::select(Gene=rowname, factor = '.') %>%
      dplyr::filter(.$Gene %in% rownames(variable_factor_matrix)) %>%
      column_to_rownames(var= "Gene")  %>%
      as.matrix()
    # Adjusted regional bulk equals to the dot product of adjusted matrix and regional factor K
    region_bulk <- (adjusted_matrix[,k] %>% as.data.frame()) * K_filtered
    # Assign expression to each spot
    each_spot <- data.frame(region_bulk/length(cluster$orig.ident))
    output <- cbind(each_spot, rep(each_spot[1],  length(cluster$orig.ident)))
    output <- output[1:(length(output)-1)]
    colnames(output) <- rownames(cluster@meta.data)
    output_region <- output_region %>%
      cbind(output)
    }
  output_final <- output_final %>%
    cbind(output_region)

### Set 500 iterations as the initial matrix
  if (i == check_interval) {
    cell_list <- as.list(rownames(slice_4@meta.data))
    spot_average_counts_final <- data.frame(matrix(ncol = 0, nrow = length(rownames(variable_factor_matrix))))
    rownames(spot_average_counts_final) <- rownames(variable_factor_matrix)
    spot_rep_times <- data.frame()
    for ( x in 1:length(cell_list)) {
      spot <- colnames(output_final) == cell_list[x]
      spot_counts <- output_final[, spot, drop = FALSE]
      spot_rep_times <- spot_rep_times %>%
        rbind(length(colnames(spot_counts)))
      spot_average_counts <-  data.frame(rowSums(spot_counts))/length(spot_counts)
      colnames(spot_average_counts) <- cell_list[x]
      spot_average_counts_final <- spot_average_counts_final %>%
        cbind(spot_average_counts)
    }
    current_matrix <- spot_average_counts_final
    current_matrix[is.na(current_matrix)] <- 0
  }
### Calculate the losses between the two matrices over 500 iterations
  if (i > check_interval & i %% check_interval == 0){
    cell_list <- as.list(rownames(slice_4@meta.data))
    spot_average_counts_final <- data.frame(matrix(ncol = 0, nrow = length(rownames(variable_factor_matrix))))
    rownames(spot_average_counts_final) <- rownames(variable_factor_matrix)
    spot_rep_times <- data.frame()
    for ( y in 1:length(cell_list)) {
      spot <- colnames(output_final) == cell_list[y]
      spot_counts <- output_final[, spot, drop = FALSE]
      spot_rep_times <- spot_rep_times %>%
        rbind(length(colnames(spot_counts)))
      spot_average_counts <-  data.frame(rowSums(spot_counts))/length(spot_counts)
      colnames(spot_average_counts) <- cell_list[y]
      spot_average_counts_final <- spot_average_counts_final %>%
        cbind(spot_average_counts)
      }
      next_matrix <- spot_average_counts_final 
      next_matrix[is.na(next_matrix)] <- 0
      losses[i] <- loss_function(current_matrix, next_matrix)
      current_matrix <- next_matrix
      if(i > 3*check_interval){
        if((abs(losses[i] - losses[i-check_interval]) < threshold) & (abs(losses[i-check_interval] - losses[i-2*check_interval]) < threshold) &  (abs(losses[i-2*check_interval] - losses[i-3*check_interval]) < threshold)){
    final_df <- spot_average_counts_final
    optimal_iteration <- i 
    break
        }
      }
  }
}
# slice_4_adjusted <- CreateSeuratObject(final_df,assay = "RNA",meta.data = metadata)
# slice_4_adjusted@images$image <- slice_4@images$image
# saveRDS(slice_4_adjusted, here::here("test","slice_4_adjusted.rds"))

```

```{r visualize loss changes along iteration times, out.width="50%", fig.align = 'center', fig.cap = "The change of losses with the increase of iteration times."}
losses_df <- data.frame(x = seq(2*check_interval, optimal_iteration, check_interval))
losses_df$y <- losses[losses_df$x]

ggplot(losses_df, aes(x = x, y = y)) +
  geom_line(color = "#9999CC") +
  geom_point(size=1) +
  xlab("Iteration times") + 
  ylab("losses") +
  theme_bw() + theme(panel.border = element_blank(), panel.grid.major = element_blank(),
panel.grid.minor = element_blank(), axis.line = element_line(colour = "black"))
```

<p>&nbsp;</p>
### Palette performance assessment
<p align="justify">We assess the Palette performance through comparing the expression patterns of Palette implemented data to original ST data and _in situ_ hybridization images. Here we show the expression patterns of marker genes CG5171 (amnioserosa) and Idgf6 (hemolymph). Compared to the original ST data (Left), the Palette-implemented data (Right) significantly increases the signal-to-noise ratio and exhibits more specific expression patterns, which are more similar to the bona fide patterns observed by _in situ_ hybridization.
</p>
```{r, out.width="70%", fig.align = 'center', fig.cap = "Comparsion of expression patterns."}
# Construct sce object for expression visualization 
metadata <-  slice_4_adjusted@meta.data
metadata$col <-  slice_4_adjusted@meta.data$col
metadata$row <-  slice_4_adjusted@meta.data$row


sce_S4 <- SingleCellExperiment(assays = list(counts = slice_4_adjusted@assays$RNA@counts, logcounts = slice_4_adjusted@assays$RNA@data),
                            colData= metadata)

sce_S4 <- spatialPreprocess(sce_S4, platform="ST",
                              n.PCs=4,n.HVGs = 2000, log.normalize=FALSE,skip.PCA=TRUE)


p1 <- featurePlot(
  sce_S4,
  "CG5171",
  assay.type = "logcounts",
  diverging = FALSE,
  low = NULL,
  high = "darkblue",
  mid = NULL,
  color = NULL,
  platform = NULL,
  is.enhanced = NULL,
)

p2 <- featurePlot(
  sce_S4,
  "Idgf6",
  assay.type = "logcounts",
  diverging = FALSE,
  low = NULL,
  high = "darkblue",
  mid = NULL,
  color = NULL,
  platform = NULL,
  is.enhanced = NULL,
)

slice_5_metadata <-  slice_5@meta.data
slice_5_metadata$col <-  slice_5_metadata$new_x %>% as.numeric() %>% as.integer()
slice_5_metadata$row <-  slice_5_metadata$new_y %>% as.numeric() %>% as.integer()


sce_S5 <- SingleCellExperiment(assays = list(counts = slice_5@assays$RNA@counts, logcounts = slice_5@assays$RNA@data),
                            colData=slice_5_metadata)

sce_S5 <- spatialPreprocess(sce_S5, platform="ST",
                              n.PCs=4,n.HVGs = 2000, log.normalize=FALSE,skip.PCA=TRUE)

p3 <- featurePlot(
  sce_S5,
  "CG5171",
  assay.type = "logcounts",
  diverging = FALSE,
  low = NULL,
  high = "darkblue",
  mid = NULL,
  color = NULL,
  platform = NULL,
  is.enhanced = NULL,
)

p4 <- featurePlot(
  sce_S5,
  "Idgf6",
  assay.type = "logcounts",
  diverging = FALSE,
  low = NULL,
  high = "darkblue",
  mid = NULL,
  color = NULL,
  platform = NULL,
  is.enhanced = NULL,
)

ggarrange(p3, p1, p4, p2, ncol = 2, nrow = 2)
```
<p>&nbsp;</p>
### Session information
```{r}
devtools::session_info(pkgs = c("loaded"))
```
