==> analysis/5. Palette implementation.Rmd <==
---
title: "5. Palette implementation"
author: "Yang"
date: '2023-02-22'
output: 
  html_document: 
    toc: yes
    toc_float: yes
    code_folding: hide
editor_options: 
  chunk_output_type: console
---
### Setup

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE)
```

```{r packages}
library(limma)
library(Seurat)
library(readr)
library(edgeR)
library(tidyverse)
library(magrittr)
library(pander)
library(ggrepel)
library(scales)
library(plyr)
library(purrr)
library(scatterplot3d)
library(SingleCellExperiment)
library(Matrix.utils)
library(reshape2)
library(ggalt)
library(BayesSpace)
library(MuSiC)
library(plyr)
library(plotly)
library(rgl)

theme_set(theme_bw())
panderOptions("big.mark", ",")
panderOptions("table.split.table", Inf)
panderOptions("table.style", "rmarkdown")
if (interactive()) setwd(here::here("analysis"))
```

### data load
```{r}
# load bulk
S10_2_dgeList <- readRDS(here::here("output","S10_2_dgeList.rds"))
# load Stereo data with adjusted coordinates 
Stereo <- readRDS(here::here("output","temp","Stereo.rds"))
# load spots with coordinates in each slice
z <- readRDS(here::here("output","temp","spots_in_slices.rds"))
# load a pseudo bulk from Stereo slices
slice_counts <- readRDS(here::here("output","temp", "slice_counts.rds"))  
```

### convert to gene ID to gene name, only keep the common genes in bulk and psenudo bulk
```{r}
bulk <- S10_2_dgeList[,7:25]
Stereo <- Stereo
# number of slices
n = 19

# Gene ID/name conversion & only keep genes that can be detected in Stereo
bulk <-  bulk$counts %>% 
  as.data.frame() %>%
  rownames_to_column() %>%
  left_join(bulk$genes %>% 
              as.data.frame() %>% 
              dplyr::select(external_gene_name) %>%
              rownames_to_column()) %>%
  dplyr::select(-rowname) %>%
  .[!(duplicated(.$external_gene_name) | duplicated(.$external_gene_name, fromLast = TRUE)),] %>%
  rownames_to_column() %>%
  dplyr::filter(.$external_gene_name %in% rownames(slice_counts)) %>%
  column_to_rownames(var= "external_gene_name") %>%
  dplyr::select(-rowname)


spatial  <- list()
slice <- list()
slice_list <- list()
slice_sum <-  data.frame(matrix(ncol = 0, nrow = length(rownames(Stereo))))
slice_sum_filtered <- list()

for ( i in 1:n){
  slice[i] <- print(rownames(z[i] %>% as.data.frame()) %>% list())
  slice_list[[i]] <- print(Stereo[, slice[[i]]])
  slice_list[[i]]@meta.data <- z[i] %>% as.data.frame()
  slice_sum <- slice_sum %>%
    cbind(colSums(t(slice_list[[i]]@assays$RNA@counts)) %>% as.data.frame() %>%
            set_colnames(i))
}

# only keep genes that can be detected in bulk
slice_sum_filtered <- slice_sum %>%
    as.data.frame() %>%
    rownames_to_column() %>%
    dplyr::filter(rowname %in% rownames(bulk)) %>%
    column_to_rownames(var= "rowname")  %>%
    as.matrix()
```


### Find out stable genes for spatial clustering
```{r find out stable genes}
slice_sum_filtered_high_expressed <- list()
bulk_high_expressed_reorder <- list()
gene_filter <- list()
gene_filter_filtered <- list()
stable_genes <- list()
for(i in 1:n){
# Filter counts > 30 as high expressed genes
    slice_sum_filtered_high_expressed[[i]] <- slice_sum_filtered[,i] %>%
      as.data.frame() %>%
      set_colnames("Counts") %>%
      dplyr::filter(.$Counts > 30)
    
    bulk_high_expressed_reorder[[i]] <- bulk[match(rownames(slice_sum_filtered_high_expressed[[i]]), rownames(bulk)),i,drop = FALSE] %>%
      as.data.frame()
    
    gene_filter[[i]] <- bulk_high_expressed_reorder[[i]]/slice_sum_filtered_high_expressed[[i]] 
    
    colnames(gene_filter[[i]]) <- "factor"

# Take the top 2000 stable genes for spatial clustering 
  gene_filter_filtered[[i]] <- gene_filter[[i]]  %>%
    .[is.finite(rowSums(.)),, drop = FALSE] 
  gene_filter_filtered[[i]]$differ <- abs(gene_filter_filtered[[i]]$factor - (mean(gene_filter_filtered[[i]]$factor)+median(gene_filter_filtered[[i]]$factor))/2)
  stable_genes[[i]] <- gene_filter_filtered[[i]] %>%
    .[order(.$differ),] %>%
    top_n(.,-2000)
}
```

### Define spot cluster (spatial clustering)
```{r}
metadata <- list()
sce <- list()
sce_spatial <- list()
p <- list()
cluster <- list()
prop.table <- data.frame(matrix(ncol = 4))
colnames(prop.table) <- seq(1,4)
for(i in 1:n){
metadata[[i]] <- slice_list[[i]]@meta.data %>%
  as.data.frame() %>%
  .[,c("spatial_x_adjusted_3","spatial_y_adjusted_3")]/15
colnames(metadata[[i]]) <- c("col", "row")
metadata[[i]]$row <-  as.integer(metadata[[i]]$row)
metadata[[i]]$col <-  as.integer(metadata[[i]]$col)

sce[[i]] <- SingleCellExperiment(assays = list(counts = slice_list[[i]]@assays$RNA@counts, logcounts = slice_list[[i]]@assays$RNA@counts),
                            colData=metadata[[i]]) %>%
  .[rownames(stable_genes[[i]]),]

sce_spatial[[i]] <- spatialPreprocess(sce[[i]], platform="ST", 
                              n.PCs=10,n.HVGs = 500, log.normalize=FALSE)
p[[i]] <- qTune(sce_spatial[[i]], qs=seq(2, 10), platform="ST")

# The spots on each slice is grouped into 4 clusters by kmeans. 
set.seed(149)
cluster[[i]] <- spatialCluster(p[[i]], q=4, platform="ST", d=7,
                           init.method="kmeans", model="t", gamma=2,
                           nrep=20000, burn.in=100,
                           save.chain=TRUE)
slice_list[[i]]$spatial_cluster <- cluster[[i]]$spatial.cluster
Idents(slice_list[[i]]) <- slice_list[[i]]$spatial_cluster

tab <- prop.table(table(slice_list[[i]]$spatial_cluster)) 


prop.table <- prop.table %>%
  rbind(tab) %>%
  as.data.frame() 

}

### Check the clustering 
clusterPlot(cluster[[9]], color="black") +
  theme_bw() +
  xlab("Column") +
  ylab("Row")
```

### Estimate abuundance of clusters in the bulk (slice deconvolution) & Generate Adjusted Matrix
```{r}
spot_numbers <- data.frame()
for(i in 1:(n-1)){
  numbers <- length(colnames(slice_list[[i]])) %>%
    as.data.frame()
  spot_numbers <- spot_numbers %>%
    rbind(numbers)
}
spot_max <- max(length(colnames(slice_list[[i]])))

sc.eset <- list() 
Est.pro <- list()
Est.pro_table <- list()
cluster.averages <- list()
cluster.averages_filtered <- list()
adjusted_matrix <- list()
bulk_reorder <- list()
variable_factor <- list()
variable_factor_filtered <-list()
variable_factor_filtered_2 <- list()
lib_size_ratio <- list()

for(i in 1:(n-1)){

bulk.est <- Biobase::ExpressionSet(assayData = (bulk %>% as.matrix()))
sc.eset[[i]] <- ExpressionSet(assayData = as.matrix(GetAssayData(slice_list[[i]][rownames(stable_genes[[i]]),])), phenoData = new("AnnotatedDataFrame", slice_list[[i]]@meta.data))

# Perform deconvolution on bulk with cluster expression data from spatial clustering as reference
slice_list[[i]]$sample_id <- colnames(slice_list[[i]])

sce[[i]] <- SingleCellExperiment(assays = list(counts = slice_list[[i]][rownames(stable_genes),]@assays$RNA@counts, logcounts = slice_list[[i]][rownames(stable_genes),]@assays$RNA@data),
                            colData=slice_list[[i]]@meta.data)

Est.pro[[i]] <- music_prop(bulk.mtx = exprs(bulk.est), sc.sce = sce[[i]],  clusters = "spatial_cluster", samples = "sample_id") 

# abundance = Est.pro_table
Est.pro_table[[i]] <- Est.pro[[i]]$Est.prop.weighted %>%
  .[i, c("1","2","3","4")] %>% as.data.frame() %>% as.matrix()
cluster.averages[[i]] <- AverageExpression(slice_list[[i]]) %>%
  as.data.frame() %>%
  as.matrix()
cluster.averages_filtered[[i]] <- cluster.averages[[i]] %>%
  as.data.frame() %>%
  rownames_to_column() %>% 
  dplyr::filter(.$rowname %in% rownames(bulk.est)) %>%
  column_to_rownames(var= "rowname")  %>%
  as.matrix() %>%
  .[,c("RNA.1","RNA.2","RNA.3","RNA.4")]


bulk_reorder[[i]] <- bulk[match(rownames(cluster.averages_filtered[[i]]), rownames(bulk)),i,drop = FALSE] %>% as.data.frame()

# Generate adjusted matrix
variable_factor[[i]]  <- bulk_reorder[[i]]/(cluster.averages_filtered[[i]]  %*% Est.pro_table[[i]]) 

variable_factor_filtered[[i]] <- variable_factor[[i]] %>% 
  .[is.finite(rowSums(.)),, drop = FALSE] %>%
    as.data.frame() 


variable_factor_filtered_2[[i]] <- rep(variable_factor_filtered[[i]],4) %>%
  as.data.frame() %>%
  set_colnames(c("1","2","3","4"))
rownames(variable_factor_filtered_2[[i]] ) <- rownames(variable_factor_filtered[[i]])

# Normalized by spot numbers
adjusted_matrix[[i]] <- (cluster.averages_filtered[[i]])[rownames(variable_factor_filtered_2[[i]]),] * variable_factor_filtered_2[[i]] * (length(colnames(slice_list[[i]]))/spot_max)
}
```

### Estimate expression of each spot
```{r message=FALSE}
slice.random.obj <- list()
slice.random.obj.neighbor <- list()
output_final_list <- list()
for (j in 1:(n-1)) {
  output_final <- data.frame(matrix(ncol = 0, nrow = length(rownames(variable_factor_filtered[[j]]))))
  rownames(output_final) <- rownames(variable_factor_filtered[[j]])
  for (i in 1:5000) {
# select one random spot and its neighboring region
    set.seed(i)
    slice.random.obj[[j]] <- slice_list[[j]][, sample(colnames(slice_list[[j]]), size = 1, replace=F)]
    spot <- slice.random.obj[[j]]@meta.data[,c("spatial_x_adjusted_3","spatial_y_adjusted_3")] %>% as.data.frame()
    df <- slice_list[[j]]@meta.data[,c("spatial_x_adjusted_3","spatial_y_adjusted_3")] %>% as.data.frame()
    df$distance = sqrt((df$spatial_x_adjusted_3 - spot$spatial_x_adjusted_3)^2 + (df$spatial_y_adjusted_3 - spot$spatial_y_adjusted_3)^2)
    df <- df[order(df$distance),][1:25,] 
    slice.random.obj.neighbor[[j]]  <-  slice_list[[j]] [,rownames(df)]
# Sum counts to construct pseudo bulk of the slice (ST)
    ST <- colSums(t((slice_list[[j]]@assays$RNA@counts))) %>%
      as.data.frame() %>%
      rownames_to_column() %>%
      dplyr::select(Gene=rowname, Counts = '.')
    
    output_region <- data.frame(matrix(ncol = 0, nrow = length(rownames(variable_factor_filtered[[j]]))))
    rownames(output_region) <- rownames(variable_factor_filtered[[j]])
    
    for (k in as.numeric(levels(as.factor(slice.random.obj.neighbor[[j]]$spatial_cluster)))){
      cluster <- subset(slice.random.obj.neighbor[[j]], subset = spatial_cluster == k)
# for each cluster in these 25 spots, do a regional cluster pseudo bulk (RST)
      RST <- colSums(t((cluster@assays$RNA@counts))) %>%
        as.data.frame() %>%
        rownames_to_column() %>%
        dplyr::select(Gene=rowname, Counts = '.')
# Calculate regional cluster factor K
      K <- RST$Counts/ST$Counts %>%
        as.data.frame()
      rownames(K) <- ST$Gene
      
      K_filtered <- K %>%
        rownames_to_column() %>%
        dplyr::select(Gene=rowname, factor = '.') %>%
        dplyr::filter(.$Gene %in% rownames(adjusted_matrix[[j]])) %>%
        column_to_rownames(var= "Gene")  %>%
        as.matrix()
      
      region_bulk <- adjusted_matrix[[j]][,k,drop = FALSE] * K_filtered
# Evenly allocating expression into each spot of the same cluster
      each_spot <- data.frame(region_bulk/length(cluster$orig.ident))
      output <- cbind(each_spot, rep(each_spot[1],  length(cluster$orig.ident)))
      output <- output[1:(length(output)-1)]
      colnames(output) <- rownames(cluster@meta.data)
      output_region <- output_region %>%
        cbind(output)
      }
    output_final <- output_final %>% 
      cbind(output_region)
    }
  output_final_list[[j]] <- output_final
  }
```

### Calculate spot average expression after cycling
```{r}
cell_list <- list()
spot_average_counts_final_list <- list()
spot_rep_times_list <- list()
for ( j in 1:(n-1)){
cell_list[[j]] <- rownames(slice_list[[j]]@meta.data)
spot_average_counts_final <- data.frame(matrix(ncol = 0, nrow = length(rownames(variable_factor_filtered[[j]]))))
rownames(spot_average_counts_final) <- rownames(variable_factor_filtered[[j]])
spot_rep_times <- data.frame()

for ( i in 1:length(cell_list[[j]])) {
  spot <- colnames(output_final_list[[j]]) == cell_list[[j]][i] 
  spot_counts <- output_final_list[[j]][, spot, drop = FALSE]
  spot_rep_times <- spot_rep_times %>%
    rbind(length(colnames(spot_counts)))
  spot_average_counts <-  data.frame(rowSums(spot_counts))/length(spot_counts)
  colnames(spot_average_counts) <- cell_list[[j]][i] 
  spot_average_counts_final <- spot_average_counts_final %>%
    cbind(spot_average_counts)
}
rownames(spot_rep_times) <- cell_list[[j]]
spot_rep_times_list[[j]] <- spot_rep_times %>%
  set_colnames("times") 
spot_average_counts_final_list[[j]] <- spot_average_counts_final
}
```

### Convert into one seurat object 
```{r}
slice_seurat <-list()
for ( i in 1:(n-1)){
  slice_seurat[[i]] <- CreateSeuratObject(counts = spot_average_counts_final_list[[i]], meta.data = slice_list[[i]]@meta.data)
}

write_rds(slice_seurat, here::here("output","temp","zf10_st_slice_seurat.rds"))

# Merge all slices
library(scCustomize)
Merged_seurat <- Merge_Seurat_List(slice_seurat,add.cell.ids = NULL,merge.data = TRUE,project = "SeuratProject")

bulk_coord.df = data.frame(x=Merged_seurat@meta.data[colnames(Merged_seurat),"spatial_x_adjusted_3",drop = FALSE], y=Merged_seurat@meta.data[colnames(Merged_seurat),"spatial_y_adjusted_3",drop = FALSE], z=Merged_seurat@meta.data[colnames(Merged_seurat),"spatial_z_adjusted_3",drop = FALSE], stringsAsFactors=FALSE) 


Merged_seurat@images$image =  new(
    Class = 'SlideSeq',
    assay = "Spatial",
    key = "image_",
    coordinates = bulk_coord.df
  )
```

### Outlier removal 
```{r}
# Regard the top 1% values as abnormal and assign the maximum value to these spots.
expression_matrix <- Merged_seurat@assays$RNA@data
expression_df <- data.frame()
for (i in rownames(expression_matrix)) {
  feature <- expression_matrix[i,] %>% as.data.frame() 
  feature_expression <- feature %>%
    set_colnames("expression") %>%
    dplyr::filter(.$expression > 0)
  feature_expression <-feature_expression$expression
  threshold <- quantile(feature_expression, 0.99) 
  feature[feature > (threshold)] <- max(feature[feature < threshold])
  feature_df <- feature %>% 
    set_colnames(i) %>% 
    t()
  expression_df <- expression_df %>% 
    rbind(feature_df)
}

Merged_seurat_new <- CreateSeuratObject(expression_df)
Merged_seurat_new@meta.data <- Merged_seurat@meta.data
Merged_seurat_new@images$image =  new(
    Class = 'SlideSeq',
    assay = "Spatial",
    key = "image_",
    coordinates = bulk_coord.df
  )

SpatialFeaturePlot(Merged_seurat_new, feature = c("tbxta","gsc","rx3","cdx4","lft1","zic1","sox17","frzb"), alpha =  c(0.1, 1),pt.size.factor = 3)
```


```{r save data}
write_rds(Merged_seurat, here::here("output","temp","zf10_st_seurat_S10_2_palette.rds"))
write_rds(Merged_seurat_new, here::here("output","temp","zf10_st_seurat_S10_2_palette_adjusted.rds"))
```
