==> analysis/9. Morphogen patterns.Rmd <==
---
title: "9. Morphogen patterns"
author: "Yang"
date: "2024-01-03"
output: 
  html_document: 
    toc: yes
    toc_float: yes
    code_folding: hide
editor_options: 
  chunk_output_type: console
---
### Setup

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE)
```

```{r}
library(Seurat)
library(ggplot2)
library(dplyr)
library(tidyverse)
library(magrittr)
library(reshape)
library(ggalt)
library(SingleCellExperiment)
library(Matrix.utils)
library(clusterProfiler)
library(org.Dr.eg.db)
library(tidyverse)
```

### Load data
```{r}
zf10_linearizeAP <- readRDS(here::here("output","temp","zf10_subset_linearizeAP.rds"))
```

### Use sliding window to visualize gene expression intensity along AP axis
```{r}
zf10_linearizeAP$normalized_AP <- (zf10_linearizeAP$pesudo_AP_x - min(zf10_linearizeAP$pesudo_AP_x))/(max(zf10_linearizeAP$pesudo_AP_x) - min(zf10_linearizeAP$pesudo_AP_x))

# Window size

s <- 0.02
df <- matrix(NA, nrow = 16514, ncol = 0)


for (i in seq(0,0.98,s)){
  if (count(between(zf10_linearizeAP$normalized_AP,i,i+s)) > 0) {
  subset <- zf10_linearizeAP[,between(zf10_linearizeAP$normalized_AP,i,i+s)]
  sce <- SingleCellExperiment(assays = list(counts = subset@assays$RNA@counts), 
                              colData = subset@meta.data)
  groups <- colData(sce)[, c("orig.ident")]
  bulk <- aggregate.Matrix(t(counts(sce)),groupings = groups, fun = "sum") %>%
  t() %>%
  as.data.frame()
  colnames(bulk) <- i+s/2
  df <- df %>% cbind(bulk)
  }
}

df_norm <- df/do.call(pmax, df)

df_subset <- df_norm[c("fgf24"),] %>% t() %>% as.data.frame() %>% rownames_to_column()
df_subset_1 <- melt(df_subset, id.vars= "rowname")
df_subset_1$rowname <- df_subset_1$rowname %>% as.numeric()
df_subset_1$value <- df_subset_1$value %>% as.numeric()

zf10_linearizeAP$normalized_AP <- zf10_linearizeAP$normalized_AP %>% as.numeric()

ggplot(df_subset_1, aes(x = rowname, y = value)) +
  ggforce::geom_link2(aes(group=variable, color = rowname), size = 1, n = 500, lineend = "round") +
  labs(x="Normalized_AP", y="Intensity" ) +
  xlim(0,1.05) + 
  scale_y_continuous("Intensity", breaks = c(0,0.5,1), labels = c(0,0.5,1), limits = c(-0.15,1)) + 
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(), panel.background = element_blank(), axis.line = element_line(colour = "black"), legend.position = "none",axis.ticks.x=element_blank(),axis.text.x=element_blank()) +
  scale_color_distiller(palette = "RdBu", direction =1)
```

### Spatial gene plot
```{r}
SpatialFeaturePlot(zf10_linearizeAP, feature = c("aldh1a2"), alpha =  c(0.1, 1),pt.size.factor = 4,image.alpha = 0,stroke = 0.01) + ggplot2::theme_classic() +
  scale_fill_gradientn(colors=alpha(c("lightgrey","darkblue"), 1)) 

ggsave("hoxb2a.tiff",path = here::here("analysis","Morphogen_images"), width = 3, height = 3, device='tiff', dpi=700)
```

### Plot ligands showing similar patterns (posterior trunk and tail)
```{r}
df_subset <- df_norm[c("aldh1a2","wnt11","wnt3a","wnt3","bmp2b","bmp4","bmp7a","gdf11","fgf8a","wnt5b","jag1a","gdf3","fgf10a","fgf24"),] %>% t() %>% as.data.frame() %>% rownames_to_column()

df_subset_1 <- melt(df_subset, id.vars= "rowname")
df_subset_1$rowname <- df_subset_1$rowname %>% as.numeric()
df_subset_1$value <- df_subset_1$value %>% as.numeric()

ggplot() +
  geom_line(data = df_subset_1, aes(rowname, y = value, col=variable), linetype="solid", size=1) +
  labs(x="Position", y="Intensity" ) +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(), panel.background = element_blank(), axis.line = element_line(colour = "black"),legend.position = "none")  +
  scale_color_manual(values = c("#801818","#FB607F","#151B54","black", "#1E90FF", "#82CAFF", "#4863A0", "#566D7E","#6F2DA8","#82EEFD","#b9bfff","#48aaad","#e4a0f7","#B200ED")) +
  geom_point(data = zf10_linearizeAP@meta.data,aes(zf10_linearizeAP$normalized_AP, runif(10227, min=-0.2, max=-0.05)),colour = zf10_linearizeAP$colors, size =0.3) +
  xlim(0,1.05) + 
  scale_y_continuous("Intensity", breaks = c(0,0.5,1), labels = c(0,0.5,1), limits = c(-0.2,1)) 
```

### Calculate cell proportion (in sliding window) and plot
```{r}
s <- 0.02
prop_all <- NULL
for (i in seq(0,0.98,s)){
  if (i != 0.92 ) {
  subset <- zf10_linearizeAP[,between(zf10_linearizeAP$normalized_AP,i,i+s)]
  prop <- prop.table(table(subset$bin_annotation)) %>% as.data.frame()
  colnames(prop) <- c("cell_type", "percentage") 
  prop$condition <- (i + s/2)
  prop_all <- prop_all %>% rbind(prop)
  }
}


cell_types <- c("Polster","Anterior Neural Keel","Posterior Neural Keel")

other_prop <- NULL

for (i in prop_all$condition) {
  prop <- data.frame(cell_type = "Other",
                           percentage = sum(prop_all[which(prop_all$condition== i & prop_all$cell_type %in% cell_types), 2]),
                           condition = i)
  other_prop <- other_prop %>% rbind(prop)
}
other_prop <- other_prop %>% .[!duplicated(.$condition),]

prop_select <- prop_all %>% dplyr::filter(.$cell_type %in% c("Paraxial Mesoderm, Neural Keel","Notochord","Segmental Plate, Tail Bud")) %>% rbind(other_prop)

prop_select$cell_type <- factor(prop_select$cell_type, levels = c("Other","Segmental Plate, Tail Bud","Notochord","Paraxial Mesoderm, Neural Keel"))

ggplot(prop_select, aes(x = condition, y = percentage, fill = cell_type))+
  geom_bar(stat = "identity") + 
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(), panel.background = element_blank(), axis.line = element_line(colour = "black"),legend.position = "none") +
  xlim(0,1.05) +
  labs(x="Normalized AP", y = "Cell proportion") +
  scale_fill_manual(values = c("grey","#6E9BC5","#FEAF16","#683B79")) 

ggplot(prop_all, aes(x = condition, y = percentage, fill = cell_type))+
   geom_bar(stat = "identity") +
  xlim(0,1.05) 
```

### Based on cell positions, select cells for comparison (GO enrichment analysis)
```{r}
zf10_linearizeAP_subset <- subset(zf10_linearizeAP, subset = bin_annotation == "Paraxial Mesoderm, Neural Keel" | bin_annotation == "Notochord" | bin_annotation == "Segmental Plate, Tail Bud")

subset_1 <- zf10_linearizeAP_subset[,between(zf10_linearizeAP_subset$normalized_AP,0.36,0.54)]
subset_2 <- zf10_linearizeAP_subset[,between(zf10_linearizeAP_subset$normalized_AP,0.54,0.80)]

subset_1$ident <- "posterior trunk"
subset_2$ident <- "tail"
subset <- merge(subset_1, subset_2) 
Idents(subset) <- "ident"

DE <- FindMarkers(subset,ident.1 = "posterior trunk", ident.2 = "tail")
somite <- DE %>% dplyr::filter(avg_log2FC > 0.3) %>%
  dplyr::filter(p_val_adj < 0.05)

DE_df <- DE %>% 
  dplyr::filter(avg_log2FC > 0.3 | avg_log2FC < -0.3) %>%
  dplyr::filter(p_val_adj < 0.05) %>%
  rownames_to_column(var = "gene_name") 


write_csv(DE_df, here::here("output","Data S1. DE genes in posterior trunk and tail (posterior trunk versus tail) at 12 hpf.csv"))

somite_ego <- enrichGO(gene         = rownames(somite),
                OrgDb         = org.Dr.eg.db,
                keyType       = 'SYMBOL',
                ont           = "BP",
                pAdjustMethod = "BH",
                pvalueCutoff  = 0.05,
                qvalueCutoff  = 0.05) 
somite_ego_simplied <-  simplify(somite_ego) 
dotplot(somite_ego_simplied, showCategory = 8) + scale_color_viridis_c()
                                 
write_csv(somite_ego_simplied@result, here::here("output","Data S2. Enriched GO in posterior trunk at 10 hpf.csv"))

somite_ego_select<- somite_ego_simplied[c("GO:0009952", "GO:0061053", "GO:0030902", "GO:0001756", "GO:0060485", "GO:0001501", "GO:0061448", "GO:0007369")] %>% as.data.frame()
somite_ego_select$GeneRatio <- somite_ego_select$Count/1011
somite_ego_select <- somite_ego_select[order(somite_ego_select$GeneRatio),]
somite_ego_select <- somite_ego_select %>% mutate(Description = fct_reorder(Description, Count))

ggplot(somite_ego_select, # you can replace the numbers to the row number of pathway of your interest
             aes(x = GeneRatio, y = Description)) + 
             geom_point(aes(size = Count, color = p.adjust)) +
             theme_bw(base_size = 10) +
             xlim(0.015,0.045) +
             ylab(NULL) +
             scale_color_viridis_c()


tail <- DE %>% dplyr::filter(avg_log2FC < -0.3) %>%
  dplyr::filter(p_val_adj < 0.05)

tail_ego <- enrichGO(gene         = rownames(tail),
                OrgDb         = org.Dr.eg.db,
                keyType       = 'SYMBOL',
                ont           = "BP",
                pAdjustMethod = "BH",
                pvalueCutoff  = 0.05,
                qvalueCutoff  = 0.05) 
tail_ego_simplied <-  simplify(tail_ego) 
dotplot(tail_ego_simplied, showCategory = 8) + scale_color_viridis_c()

write_csv(tail_ego_simplied@result, here::here("output","Data S3. Enriched GO in tail region at 10 hpf.csv"))

tail_ego_select<- tail_ego_simplied[c("GO:0061053", "GO:0009952", "GO:0001756", "GO:0048332", "GO:0048705", "GO:0007492", "GO:0009890", "GO:0009799")] %>% as.data.frame()
tail_ego_select$GeneRatio <- tail_ego_select$Count/2108
tail_ego_select <- tail_ego_select[order(tail_ego_select$GeneRatio),]
tail_ego_select <- tail_ego_select %>% mutate(Description = fct_reorder(Description, Count))


ggplot(tail_ego_select, # you can replace the numbers to the row number of pathway of your interest
             aes(x = GeneRatio, y = Description)) + 
             geom_point(aes(size = Count, color = p.adjust)) +
             theme_bw(base_size = 10) +
             xlim(0.008,0.043) +
             ylab(NULL) +
             scale_color_viridis_c()
```

### Use a small sliding window to generate more data points for construct model (machine learning)
```{r}
# Window size
s <- 0.001
df <- matrix(NA, nrow = 16514, ncol = 0)


for (i in seq(0,0.999,s)){
  if (count(between(zf10_linearizeAP$normalized_AP,i,i+s)) > 0) {
  subset <- zf10_linearizeAP[,between(zf10_linearizeAP$normalized_AP,i,i+s)]
  sce <- SingleCellExperiment(assays = list(counts = subset@assays$RNA@counts), 
                              colData = subset@meta.data)
  groups <- colData(sce)[, c("orig.ident")]
  bulk <- aggregate.Matrix(t(counts(sce)),groupings = groups, fun = "sum") %>%
  t() %>%
  as.data.frame()
  colnames(bulk) <- i+s/2
  df <- df %>% cbind(bulk)
  }
}

df_norm <- df/do.call(pmax, df)
```

### Take all ligand and TF expression
```{r}
df_ligands <- df_norm[c("aldh1a2","wnt5b","wnt11","wnt16","jag1a","jag1b","jag2b","shha", "wnt3a","wnt3", "wnt8b","bmp1a","bmp1b","bmp2b","bmp4","bmp7a","bmp7b","gdf3","gdf6a","gdf11","fgf3", "fgf8a", "fgf10a", "fgf17","fgf18b","fgf24"),] %>% t() %>% as.data.frame() %>% rownames_to_column()

saveRDS(df_ligands, here::here("output","temp","zf10_ligands.rds"))

TF <- read.table(here::here("reference","Danio_rerio_TF.txt"),sep="\t",header = T) %>% as.data.frame() 
TF_select <- TF[TF$Symbol %in% rownames(df_norm),]


df_TF <- df_norm[TF_select$Symbol,] %>% na.omit() %>% t() %>% as.data.frame() %>% rownames_to_column() 
saveRDS(df_TF, here::here("output","temp","zf10_TF.rds"))
```
